using Yarn 4.7.0 with nodeLinker: node-modules

+++ clean-modules.bat
@@ -1,19 +1,22 @@
 @echo off
 
-echo Removing node_modules...
-rmdir /s /q node_modules
+echo Removing .yarn folder...
+if exist .yarn\cache rmdir /s /q .yarn\cache
+
+echo Removing node_modules folder...
+if exist node_modules rmdir /s /q node_modules
 
 echo Removing .expo...
-rmdir /s /q .expo
+if exist .expo rmdir /s /q .expo
 
 echo Removing android...
-rmdir /s /q android
+if exist android\build rmdir /s /q android\build
 
-echo Deleting package-lock.json...
-del package-lock.json
+echo Deleting yarn.lock...
+if exist yarn.lock del yarn.lock
 
-echo clean npm...
-npm cache clean --force
+echo clean cache...
+yarn cache clean --all
 
-echo Done , you can run npm install now.
+echo Done. Run "yarn" to reinstall dependencies.
 pause
